portfolio/requirements.txt
Flask==2.2.5
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
gunicorn==21.2.0
